pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "little-alchemy-2-text-mcp"
version = "1.0.0"
description = "Little Alchemy 2 Text game playable via Model Context Protocol (MCP)"
authors = [
    { name = "Little Alchemy 2 Text MCP", email = "example@example.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
keywords = ["mcp", "little-alchemy", "game", "text-based", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "gymnasium>=1.0.0",
    "numpy>=1.20.0",
    "python-box>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/your-username/little-alchemy-2-text-mcp"
Repository = "https://github.com/your-username/little-alchemy-2-text-mcp"
Issues = "https://github.com/your-username/little-alchemy-2-text-mcp/issues"

[project.scripts]
little-alchemy-2-text-mcp = "mcp_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["env*", "players*", "utils*"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "gymnasium.*",
    "box.*",
    "numpy.*",
]
ignore_missing_imports = true